requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via low-latency-stream-kit (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
cbor2==5.7.0
    # via low-latency-stream-kit (pyproject.toml)
click==8.3.0
    # via typer
dpkt==1.9.8
    # via low-latency-stream-kit (pyproject.toml)
jsonschema==4.25.1
    # via low-latency-stream-kit (pyproject.toml)
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
nats-py==2.11.0
    # via low-latency-stream-kit (pyproject.toml)
pydantic==2.11.9
    # via low-latency-stream-kit (pyproject.toml)
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==14.1.0
    # via typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
typer==0.19.2
    # via low-latency-stream-kit (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
PyQt5==5.15.11
pytest-qt==4.4.0
